+++ src/main/java/org/chzzk/howmeet/domain/regular/auth/dto/login/response/MemberLoginResponse.java
@@ -1,6 +1,6 @@
 package org.chzzk.howmeet.domain.regular.auth.dto.login.response;
 
-import org.chzzk.howmeet.domain.regular.auth.entity.Member;
+import org.chzzk.howmeet.domain.regular.member.entity.Member;
 
 public record MemberLoginResponse(String accessToken, Long memberId, String nickname) {
     public static MemberLoginResponse of(final String accessToken, final Member member) {
