+++ src/main/java/org/chzzk/howmeet/domain/regular/member/entity/Member.java
@@ -1,4 +1,4 @@
-package org.chzzk.howmeet.domain.regular.auth.entity;
+package org.chzzk.howmeet.domain.regular.member.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
