+++ src/main/java/org/chzzk/howmeet/domain/regular/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package org.chzzk.howmeet.domain.regular.member.repository;
 
-import org.chzzk.howmeet.domain.regular.auth.entity.Member;
+import org.chzzk.howmeet.domain.regular.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
